requirements.txt
alipay==0.7.4
amqp==2.1.4
appdirs==1.4.2
asn1crypto==0.22.0
attrs==16.3.0
Automat==0.5.0
Babel==2.4.0
bcrypt==3.1.3
beautifulsoup4==4.5.3
billiard==3.5.0.2
bottlenose==1.1.2
bs4==0.0.1
celery==4.0.2
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
contextlib2==0.5.5
cryptography==1.8.1
cssselect==1.0.1
Django==1.10
django-admin-honeypot==1.0.0
django-appconf==1.0.2
django-axes==2.3.2
django-braces==1.11.0
django-celery-results==1.0.1
django-ckeditor==5.2.2
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-daterange-filter==1.3.0
django-debug-toolbar==1.7
django-flat-responsive==1.3.0
django-rest==0.0.1
django-taggit==0.22.0
django-tinymce==2.6.0
django-wysiwyg-redactor==0.5.1
djangorestframework==3.5.4
docopt==0.4.0
EasyProcess==0.2.3
flower==0.9.1
gevent==1.2.2
greenlet==0.4.12
idna==2.5
incremental==16.10.1
itsdangerous==0.24
kombu==4.0.2
lxml==3.7.3
mailchimp==2.0.9
Markdown==2.6.8
mysqlclient==1.3.10
numpy==1.12.0
olefile==0.44
packaging==16.8
pandas==0.19.2
paramiko==2.2.1
parsel==1.1.0
Pillow==4.1.0
psutil==5.2.2
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
PyDispatcher==2.0.5
PyNaCl==1.1.2
pyOpenSSL==16.2.0
pyparsing==2.1.10
pytesseract==0.1.6
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.0
pytz==2016.10
PyVirtualDisplay==0.2.1
queuelib==1.4.2
raven==6.0.0
rcssmin==1.0.6
redis==2.10.5
requests==2.13.0
rjsmin==1.0.12
Scrapy==1.3.3
selenium==3.3.3
service-identity==16.0.0
six==1.10.0
SQLAlchemy==1.1.10
sqlparse==0.2.3
tornado==4.2
Twisted==17.1.0
uWSGI==2.0.15
vine==1.1.3
w3lib==1.17.0
web.py==0.40.dev0
zope.interface==4.3.3
